organized lib folder a bit
