+++ src/lib/airtable/schema.js
