+++ src/screens/general/GeneralOwnerDashboard.js
@@ -6,7 +6,7 @@ import {
   getOwnerById,
   getProjectGroupById,
   getSolarProjectById
-} from '../../lib/request';
+} from '../../lib/airtable/request';
 import { getLoggedInUserId, logOut } from '../../lib/auth';
 import AnnouncementList from '../../components/AnnouncementList';
 import LoadingComponent from '../../components/LoadingComponent';
