+++ src/screens/onboarding/Payment.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { PayPalButton } from 'react-paypal-button-v2';
 import { recordShareBuySuccess } from '../../lib/paypal';
-import { updatePerson } from '../../lib/request';
+import { updatePerson } from '../../lib/airtable/request';
 
 const clientId = process.env.REACT_APP_PAYPAL_CLIENT_ID;
